pytest: Separate Basdai tests
